detector/requirements.txt
pyyaml
psutil
flask
requests
psutil
